ai-recon-agent/pyproject.toml
[tool.setuptools.packages.find]
include = ["agent*"]

[project]
name = "ai-recon-agent"
version = "0.1.0"
description = "Autonomous AI recon agent powered by LangChain"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",      # or langchain-anthropic
    "langchain-community>=0.2.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.6.0",
    "python-whois>=0.9.0",
    "requests>=2.31.0",
    "rich>=13.0.0",                  # beautiful CLI output
    "pydantic>=2.0.0",
    "python-nmap>=0.7.1",            # nmap wrapper (nmap must be installed)
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "black", "ruff"]

[project.scripts]
recon = "main:main"
